backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.3
pydantic-settings==2.10.1
sqlalchemy==2.0.40
httpx==0.28.1
numpy==2.2.4
faiss-cpu==1.10.0
sentence-transformers==4.0.2
transformers==4.51.0
duckdb==1.2.2
beautifulsoup4==4.13.4
lxml==5.4.0
geopy==2.4.1
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
pillow==11.1.0
pytesseract==0.3.13
openai-whisper==20240930
pytest==8.3.5
